Push the iterator one level further into the mapper.
This isn't a meaningful change yet, but makes it clearer how we can turn
hwt's mapper into a real iterator (at the moment we map the entire trace
and then expose an iterator).
